Converts EuiFieldNumber to Typescript
